Add redirect from als to gsw
The 'als' is used in a non-standard way in MediaWiki -
it may be used to represent the Allemanic language,
the standard code of which is 'gsw', while 'als'
is ISO 639 3 refers to Tosk Albanian, which is
not currently used in any way in MediaWiki.
This local fix adds a redirect for it.

Also add a test to check that it works correctly.

Change-Id: Id904cab129eb58f8b96ce493e77d21da7c44ea8b