Preferences tests
This is a basic test that checks the writing preferences
does not result in an error.
It depends on I8fdc1a05 and I18a5ffb5.
It is very rudimentary: The saving test only checks that the API call
didn't result in an error. It should be enough to check for bug 41763,
but it doesn't really check that the options were written correctly.
It also doesn't do anything to distinguish anonymous and logged-in
users, the preferences for which work differently.

Change-Id: I417089eea64e8a84bd38a00bbb31cbb77dc5bd68