Update the $wgULSNoWebfontsSelectors defaults
$wgULSNoWebfontsSelectors had default value as '#p-lang li > a'
-the interlanguage links. It does not make sense after the class
'autonym' was added to such links. That 'autonym' css class has
fontfamily definition as Autonym, sans-serif

ie, Autonym font is applied to the interlanguage links.

So, now it make sense to have default value as .autonym to
exclude from parsing.

This also fixes the problems like Bug 58374 where in random cases
the default font for the language get applied even if it has 'autonym' class.

This issue was because the js get executed before the css for .autonym class is
ready and applied. The jquery.webfonts parser will miss to detect the explicit
fontfamily 'Autonym, sans-serif' and the default font is applied.

Bug: 58374
Change-Id: I7478a58583d3b029ae8e2d63e118917499c29c30