Remove -moz-linear-gradient
This does not have much effect on the gzipped payload, but it still
makes the modules look big and (I assume) take space in the local
storage.

This prefix is not used by current versions of browsers, so the only
change is that some users with old browsers get bitmap images instead
of vector images.

Instances where -moz-linear-gradient is used with colors were left in,
as were the instances in upstream libs.

Change-Id: I1515c53955467eb21c8ed70e46aecc61ab4718c2