Wrong border color for caret
After seeing this wrong color dozens of times while testing svg
icons on high zooms, I can't stand the slightly different color of
the border for the caret from the rest of the ULS dialog. The border
of ULS is transparent with 0.2 alpha component over the background,
in this case I just hardcoded the resulting color here to match.

Change-Id: Ib2b2801f51ea02d2788d8703e3410eca9b3a57a3