Reduce line-height of possibly broken "200 more languages" label
Depending on the browser, skin, zoom and language the "200 more
languages" label breaks into two lines. The line height is so big
that it looks like "200 weitere

Sprachen" in German. The margin to the top (where the [...] button
is) and to the bottom (where the Wikidata "Edit links" link is) is
much smaller than the line-height.

My additional margin is choosen in a way that the design is still
pixel perfect if the text fits in one line.

Change-Id: I2fd797998a4f4e72a3764a72873f1356efc0dfb3