Move all the messages from i18n.php to json
This will require the import of all the translations from
the ext-universallanguageselector group
to the ext-jquery-uls group.

The messages already exists, so they are not re-imported.
In the next sync when the translatewiki config is updated,
the json message group of ULS will claim ownership of these messages.

Full export is needed to populate the messages in the existing json
files.

Change-Id: I6c5de42784971f59d1ca6b975095137a64c87363