Remove !defined( 'MEDIAWIKI' )
This may be causing a problem for the translate extension using
composer.json to run php code sniffer.

And not needed since we use extension.json if we use mediawiki 1.25+.

I traced back to the extension since translate extension in composer
requires this extension and seems to have that message.

Change-Id: I6c119c8f4744623e8756ce8da1869dfd66a3c026