Set a relative size for the pt-uls icon and padding so it scales properly
Instead of background-size:containing to the relative font size, but with
fixed horizontal space for it due to the padding, set both the padding and
background-size to use ems instead of px so both scale properly.

Tested in modern, cologneblue, monobook, and vector with no visible issues.
Resolves issue with icon being too large on timeless. Does not appear to
affect minerva, as that just handles it all another way.

Change-Id: Ibcea827fe15ee50c947e9c0de1e6cf053e299033