Compact language list preference should be ignored for Vector 2022 skin
The Compact language list preference was intended to control whether to
show all languages directly or show just a few of them (with option to
search for the rest) instead. However, with the new Vector 2022 skin,
there is only a button to access language selection and thus, this
setting should be ignored for this skin.

This patch modifies the "isCompactLinksEnabled" method inside Hooks
to always return true when the skin is set to Vector 2022 (when ULS
features are enabled).

Bug: T319690
Change-Id: Ic8a537d31ee4dad02932a04017f19537140e0c1c