Remove all TTF font files
The TTF files are not served or used for anything else than converting
them to woff and woff2 formats, so there is no need to store them in the
git repo. In case a new format comes along, we would need to fetch these
from the git history or from upstream sources.

Change-Id: Iedc9162011b3b37587a2a176949afde6878d122d