Update language search index
The changes are mostly automatic and add languages
that were recently added to language-data
or updated in the CLDR.

I went over the whole long diff and couldn't find
any issues that would affect the usage of the ULS
search box.

Tests are updated to reflect the current names,
but here, too, the language finding functionality
is not supposed to be affected.

Two kinds of tests were affected:
* The name of Hindi (Latin) in Malayalam: I removed
  the test because Hindi (Latin) is not used in
  MediaWiki anyway, so it doesn't matter.
* Some names of Chinese varieties. I updated
  the tests and also added aliases for languages
  whose names were changed, so that searching
  for "chinese" would find them (so it's good
  that we had those tests!).

Change-Id: I22344dadb0b01d7704ab7d76271ab27077daedb3