Add Hebrew search aliases for Tigre and Tigrinya
The Tigre and Tigrinya languages are spoken by thousands
of people in Israel, who have a more convenient access
to Hebrew keyboards than to Ethiopic keyboards.
The situation is further complicated by the fact that
the languages' names in the Hebrew alphabet have several
variant spellings, especially with regard to the first letter,
which is the most important one for searching.

This patch ensures that the languages are conveniently
findable in any Hebrew-script spelling.

(Every variant is on a separate line because
otherwise code editors mess up RTL text display.)

Bug: T375052
Change-Id: I8a6414718802f091ca1df9367b2dbc170cd568c4