Prevent language settings dialog from scaling with text size
The language settings dialog contains Codex component that,
as of Codex v2.0.0-rc.1, will scale with the Vector 2022 text
size options. To prevent uneven scaling in the dialog, this
patch adds a class that will exclude the dialog and its contents
from scaling. The class is a temporary workaround; in the future
we will either intentionally scale only the content section and
remove this workaround, create a more permanent way to exclude
certain features from scaling, or intentionally scale the
entire dialog in a consistent way.

Bug: T393650
Change-Id: I04af554596af9cb00ba3b6998f4f8f67aac7da9a