ApiLanguageSearch: Move to MediaWiki Core
We are moving the ApiLanguageSearch module from the
UniversalLanguageSelector extension to MediaWiki Core to make it
available for other extensions.

Add backward compatibility for the ApiLanguageSearch module in the
UniversalLanguageSelector extension.

LanguageNameIndexer.php, LanguageNameSearchData.php has also been ported
to MediaWiki Core as they are dependencies for the ApiLanguageSearch
module.

Bug: T408821
Change-Id: I241ef8b27e10080443673324d99b5fff3a7629e7